fix: dns setup error

+++ INSTALL.sh
@@ -134,6 +134,11 @@ install_yay() {
 }
 
 install_eww() {
+    if command -v eww >/dev/null; then
+        echo "Eww is already installed. Skipping build..."
+        return 0
+    fi
+
     run_network "Cloning Eww Repository" "cd $HOME && [ -d eww ] && rm -rf eww; git clone https://github.com/elkowar/eww"
 
     run_loader "Compiling Eww (Cargo Build)" "cd $HOME/eww && cargo build --release --no-default-features --features=wayland"
@@ -158,7 +163,7 @@ setup_network() {
 
     run_loader "Cleaning existing resolv.conf" "echo '$PASSWORD' | sudo -S rm -f /etc/resolv.conf"
 
-    run_loader "Configuring DNS (8.8.8.8 & 1.1.1.1)" "printf 'nameserver 8.8.8.8\nnameserver 1.1.1.1\n' | echo '$PASSWORD' | sudo -S tee /etc/resolv.conf"
+    run_loader "Configuring DNS" 'echo "$PASSWORD" | sudo -S sh -c "echo nameserver 8.8.8.8 > /etc/resolv.conf && echo nameserver 1.1.1.1 >> /etc/resolv.conf"'
 
     run_loader "Locking DNS settings (chattr +i)" "echo '$PASSWORD' | sudo -S chattr +i /etc/resolv.conf"
